Goodbye splash (#1057)
