+++ packages/i18n/locales/en/translation.json
@@ -697,27 +697,6 @@
     "pending": "Pending",
     "yes": "Yes"
   },
-  "splash": {
-    "anyoneCanParticipate": "Anyone from your community can be a full participant.",
-    "createExploreJoin": "Create, explore, and join DAOs",
-    "cta": "Enter the app",
-    "daosCreated": "DAOs Created",
-    "documentation": "Documentation",
-    "easyToUse": "Easy-to-use interface",
-    "easyToUseExplanation": "DAO DAO's visual interface makes governance accessible to everyone.",
-    "joinTheCommunity": "Join the community",
-    "longTagline": "Simple, capable, and free DAO tooling. Built with love, by DAO DAO, on Juno.",
-    "poweredByJuno": "Powered by Juno",
-    "proposalsCreated": "Proposals Created",
-    "proposeAndVote": "Propose & Vote",
-    "readMore": "Read more",
-    "shortTagline": "DAOs for everyone.",
-    "transparencyExplanation": "Anyone can see what decisions your DAO made, and who voted for and against them.",
-    "transparentGovernanceOnChain": "Transparent governance, entirely on-chain",
-    "usdcTotalValue": "USDC Total Value in DAOs",
-    "whatIsADao": "What is a DAO?",
-    "whatIsADaoExplanation": "A DAO is an organization democratically controlled by its members."
-  },
   "success": {
     "addedToken": "Added token to Keplr.",
     "compensationCycleCompleted": "Compensation cycle completed. Redirecting to proposal...",
