+++ packages/i18n/locales/ko/translation.json
@@ -303,27 +303,6 @@
     "yourBalance": "현재 잔액",
     "yourTokensUnstaked": "귀하의 ${{tokenSymbol}} 토큰이 언스테이킹 되었습니다."
   },
-  "splash": {
-    "anyoneCanParticipate": "커뮤니티의 모든 사용자는 참가자가 될 수 있습니다.",
-    "createExploreJoin": "DAO를 생성하고, 탐색하고, 참여하세요.",
-    "cta": "앱에 입장합니다",
-    "daosCreated": "생성 완료된 DAO",
-    "documentation": "문서",
-    "easyToUse": "사용하기 쉬운 인터페이스",
-    "easyToUseExplanation": "DAO에 투표하기 위해 더 이상 코드가 필요하지 않습니다. DAO DAO의 시각적 UI는 모든 사람이 집단적 소유권에 접근하기 용이하게 설계되었습니다.",
-    "joinTheCommunity": "커뮤니티에 가입하기",
-    "longTagline": "조작법이 단순하고 뛰어난 성능을 가진 DAO 툴링입니다. 무료로 제공되며, Juno 체인 상에서 구현되었습니다.",
-    "poweredByJuno": "Powered by Juno",
-    "proposalsCreated": "생성된 제안",
-    "proposeAndVote": "제안 & 투표",
-    "readMore": "더 읽기",
-    "shortTagline": "모두를 위한 DAO",
-    "transparencyExplanation": "누구나 귀하의 DAO가 어떤 결정을 내렸는지, 그리고 누가 찬성하고 반대했는지 알 수 있습니다. ",
-    "transparentGovernanceOnChain": "온체인 상의 투명한 거버넌스",
-    "usdcTotalValue": "DAO에 속한 USDC의 총량",
-    "whatIsADao": "DAO가 무엇인가요?",
-    "whatIsADaoExplanation": "DAO는 구성원들에 의해 민주적으로 통제되는 조직입니다."
-  },
   "success": {
     "addedToken": "Keplr에 토큰이 추가되었습니다.",
     "daoCreatedPleaseWait": "DAO 생성 완료. DAO 페이지가 준비되면 자동으로 연결됩니다...",
