+++ packages/i18n/locales/zh-tw/translation.json
@@ -325,27 +325,6 @@
   "proposalModuleLabel": {
     "DaoProposalSingle": "單項選擇"
   },
-  "splash": {
-    "anyoneCanParticipate": "任何社群中的人都可以成為正式參與者",
-    "createExploreJoin": "創立、探索和加入 DAO",
-    "cta": "進入此應用(App)",
-    "daosCreated": "已建立 DAO 總數",
-    "documentation": "文檔",
-    "easyToUse": "簡單實用的介面",
-    "easyToUseExplanation": "DAO DAO 的介面令議政對每人來說都是伸手可及",
-    "joinTheCommunity": "加入社群",
-    "longTagline": "簡單、實用和免費的 DAO 工具，由 DAO DAO 在Juno鏈上用愛建造。",
-    "poweredByJuno": "由 Juno 提供支援",
-    "proposalsCreated": "已被提交議案總數",
-    "proposeAndVote": "提交議案 & 投票",
-    "readMore": "了解更多",
-    "shortTagline": "適合所有人的 DAO",
-    "transparencyExplanation": "任何人都可以查看你的 DAO 的所有決定，還有誰支持及反對這些決定。",
-    "transparentGovernanceOnChain": "透明、紀錄在區塊鏈上的議政",
-    "usdcTotalValue": "所有 DAO 中鎖定資產轉換成 USDC 的總額",
-    "whatIsADao": "DAO 是什麼？",
-    "whatIsADaoExplanation": "DAO 是一個由其成員民主地控制的機構。"
-  },
   "success": {
     "addedToken": "該代幣已加入 Keplr",
     "daoCreatedPleaseWait": "DAO 已被建立，當 DAO 頁面準備好後你會被自動跳轉…",
