+++ packages/stateful/components/dao/index.tsx
@@ -8,5 +8,4 @@ export * from './DaoPageWrapper'
 export * from './DaoTokenDepositModal'
 export * from './DaoTreasuryHistory'
 export * from './DiscordNotifierConfigureModal'
-export * from './SplashDaoCard'
 export * from './TokenCard'
