+++ packages/stateless/components/error/ErrorBoundary.tsx
@@ -45,7 +45,7 @@ class ErrorBoundaryInner extends Component<
         <p>
           {this.props.i18n?.t?.('error.checkInternetOrTryAgain') ??
             'Check your internet connection or try again later.'}{' '}
-          <Link href="/home">
+          <Link href="/">
             <a className="underline hover:no-underline">
               {this.props.i18n?.t?.('info.considerReturningHome') ??
                 'Consider returning home.'}
