+++ packages/stateless/components/index.tsx
@@ -13,7 +13,6 @@ export * from './not_found'
 export * from './popup'
 export * from './profile'
 export * from './proposal'
-export * from './splash'
 export * from './toasts'
 export * from './token'
 export * from './tooltip'
