--- packages/stateless/components/splash/SplashCards.stories.tsx
@@ -1,13 +1,0 @@
-import { ComponentMeta, ComponentStory } from '@storybook/react'
-
-import { SplashCards } from './SplashCards'
-
-export default {
-  title: 'DAO DAO / packages / stateless / components / splash / SplashCards',
-  component: SplashCards,
-} as ComponentMeta<typeof SplashCards>
-
-const Template: ComponentStory<typeof SplashCards> = (_args) => <SplashCards />
-
-export const Default = Template.bind({})
-Default.args = {}
