--- packages/stateless/components/splash/SplashGradientWrapper.stories.tsx
@@ -1,18 +1,0 @@
-import { ComponentMeta, ComponentStory } from '@storybook/react'
-
-import { SplashGradientWrapper } from './SplashGradientWrapper'
-
-export default {
-  title:
-    'DAO DAO / packages / stateless / components / splash / SplashGradientWrapper',
-  component: SplashGradientWrapper,
-} as ComponentMeta<typeof SplashGradientWrapper>
-
-const Template: ComponentStory<typeof SplashGradientWrapper> = (args) => (
-  <SplashGradientWrapper {...args} />
-)
-
-export const Default = Template.bind({})
-Default.args = {
-  children: 'Some content',
-}
