--- packages/stateless/components/splash/index.ts
@@ -1,6 +1,0 @@
-export * from './SplashAnnouncementCard'
-export * from './SplashCards'
-export * from './SplashEnterAppButton'
-export * from './SplashGradientWrapper'
-export * from './SplashStatsCard'
-export * from './SplashVoteCard'
