+++ packages/stateless/pages/Home.stories.tsx
@@ -68,7 +68,7 @@ Connected.parameters = {
     url: 'https://www.figma.com/file/ZnQ4SMv8UUgKDZsR5YjVGH/DAO-DAO-2.0?node-id=272%3A64674',
   },
   nextRouter: {
-    asPath: '/home',
+    asPath: '/',
   },
 }
 Connected.decorators = [makeAppLayoutDecorator()]
