+++ packages/stateless/pages/Proposal.tsx
@@ -69,7 +69,7 @@ export const Proposal = ({
       <PageHeader
         breadcrumbs={{
           crumbs: [
-            { href: '/home', label: 'Home' },
+            { href: '/', label: 'Home' },
             ...getParentDaoBreadcrumbs(daoInfo.parentDao),
             { href: `/dao/${daoInfo.coreAddress}`, label: daoInfo.name },
           ],
